+++ packages/create-auto-app/CHANGELOG.md
@@ -1,5 +1,12 @@
 # create-auto-app
 
+## 0.13.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @auto-engineer/id@0.13.2
+
 ## 0.13.1
 
 ### Patch Changes