+++ packages/file-store/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @auto-engineer/file-store
 
+## 0.13.2
+
 ## 0.13.1
 
 ## 0.13.0