+++ packages/message-store/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @auto-engineer/message-store
 
+## 0.13.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @auto-engineer/message-bus@0.13.2
+
 ## 0.13.1
 
 ### Patch Changes