+++ packages/server-generator-apollo-emmett/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @auto-engineer/server-generator-apollo-emmett
 
+## 0.13.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @auto-engineer/narrative@0.13.2
+  - @auto-engineer/message-bus@0.13.2
+
 ## 0.13.1
 
 ### Patch Changes